Use uint for inputs

+++ RapidDataCodingChallenge.MathLib/Internal/IterativeMathLib.cs
@@ -9,15 +9,15 @@ using System.Threading.Tasks;
 
 internal class IterativeMathLib : IInternalMathLib
 {
-    public long Factorial(int n)
+    public long Factorial(uint n)
     {
-        return Enumerable.Range(1, n)
+        return Enumerable.Range(1, (int)n)
             .Aggregate(1L, (product, next) => product * next);
     }
 
-    public long UnevenFactorial(int n)
+    public long UnevenFactorial(uint n)
     {
-        return Enumerable.Range(1, n)
+        return Enumerable.Range(1, (int)n)
             .Where(x => x % 2 is 1)
             .Aggregate(1L, (product, next) => product * next);
     }
